Allow the location of "llvm-config" to be overridden on the "make" command line.
This commit adds an LLVM_CONFIG variable to the Makefile, and uses it rather
than directly referencing llvm-config. This allows the value to be changed on
the make command line. For example, on my system where I have installed the llvm
snapshot package for debian, I am running "make LLVM_CONFIG=llvm-config-3.4".